replaced the implicit `identity` coercion by `conforms`, which can be used to encode generalised constraints the introduction of `conforms` revealed a bug in adaptToMember, which was inferring views while already inferring one, which gave rise to diverging implicits. Predef.identity is no longer special as far as the compiler is concerned.
because conforms/identity was no longer prevented from being used as a view (which does not make sense, but preventing it shouldn't be necessary),
removeNames in NamesDefaults suddenly didn't detect all ambiguities because it relied on tryTypedApply failing
fixed by using an EmptyTree as an ambiguous argument instead of the argument, so failure is guaranteed

fixed check file for t0590

new starr

fixed the weirdest bug ever: don't know why, but can't change the total number of calls to newTermName in StdNames (so take away the one for "identity", give one back, doesn't matter where --> see "utterweirdness" at the end)
the problem manifested itself by not finding Nil. This only happens during start up (when the scala/package.scala file hasn't been compiled yet), when Nil is required before List (because that would have forced Nil to be loaded).

git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@18537 5e8d7ff9-d8ef-0310-90f0-a4852d11357a